# eventconnect-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventConnect API",
        "description": "API complète pour la gestion d'événements avec système de réservation",
        "contact": {
            "name": "EventConnect Team",
            "email": "support@eventconnect.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/upload/image": {
            "post": {
                "tags": [
                    "Fichiers"
                ],
                "summary": "Upload d'une image",
                "description": "Upload et redimensionne une image pour un événement ou un profil",
                "operationId": "uploadImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Fichier image à uploader",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "description": "Type d'image",
                                        "type": "string",
                                        "default": "event",
                                        "enum": [
                                            "event",
                                            "profile",
                                            "banner"
                                        ]
                                    },
                                    "event_id": {
                                        "description": "ID de l'événement (si type=event)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploadée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image uploadée avec succès"
                                        },
                                        "file": {
                                            "properties": {
                                                "filename": {
                                                    "type": "string"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "url": {
                                                    "type": "string"
                                                },
                                                "size": {
                                                    "type": "integer"
                                                },
                                                "mime_type": {
                                                    "type": "string"
                                                },
                                                "dimensions": {
                                                    "properties": {
                                                        "width": {
                                                            "type": "integer"
                                                        },
                                                        "height": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation du fichier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le fichier n'est pas valide"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "413": {
                        "description": "Fichier trop volumineux"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/files/{filename}": {
            "delete": {
                "tags": [
                    "Fichiers"
                ],
                "summary": "Supprimer un fichier",
                "description": "Supprime un fichier du serveur",
                "operationId": "deleteFile",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Nom du fichier à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de fichier",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "event",
                                "profile",
                                "banner"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fichier supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}/files": {
            "get": {
                "tags": [
                    "Fichiers"
                ],
                "summary": "Liste des fichiers d'un événement",
                "description": "Récupère la liste des fichiers associés à un événement",
                "operationId": "getEventFiles",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des fichiers récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}